// ts/Core/Renderer/PolygonBoxObject.d.ts
/* *
 *
 *  (c) 2016-2021 Highsoft AS
 *  Authors: Jon Arild Nygard
 *
 *  License: www.highcharts.com/license
 *
 *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
 *
 * */

export interface PolygonBoxObject {
    bottom: number;
    left: number;
    right: number;
    top: number;
}

export default PolygonBoxObject;